requirements.txt
psycopg2==2.9.9 #Postgres connection
flask==3.0.2
Gunicorn

openai
python-dotenv

requests